refactor(server): split handlers into per-group modules with a registry
Part of #132.

server.ex was a 660-line module of 26 handle/2 clauses; clause order
was load-bearing (the two resolve_choice forms disambiguated by
pattern), section comments had drifted (the scoped resolve_choice sat
under "--- Inventory ---", 227 lines from its progression sibling),
and every new command grew the pile.

Handlers now live in per-group modules — Handlers.Dice, .Systems,
.Characters, .Build, .Inventory — and server.ex is dispatch only: the
loop, the rescue boundary, protocol envelope construction, and an
explicit command -> module registry. The registry replaces
clause-order dispatch, so "unknown command" is decided by map lookup
rather than by falling off the end of 26 patterns, and each handler
module carries its own catch-all for a known command with malformed
arguments ("invalid arguments for command: ..." — previously such
requests fell through to the misleading "unknown command" response).

Shared request/response helpers (parse_display_mode,
character_with_choices_response, fetch_pending!,
validate_concept_selection!) move to Server.Common. Both
resolve_choice forms now sit adjacent in Handlers.Characters.

Handler bodies are moved verbatim aside from Common./Errors.
qualification; no behavior change except the malformed-arguments
message noted above (no test asserted on the old text).
